refactor: step3 피드백 적용
- 메소드 이동
